# docs/swagger.yaml
definitions:
  app.Card:
    properties:
      id:
        type: string
      img:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  app.CardRequest:
    properties:
      img:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  app.imageResponse:
    properties:
      url:
        type: string
    type: object
  app.orderRequest:
    properties:
      cards:
        items:
          $ref: '#/definitions/app.Card'
        type: array
    type: object
  app.orderResponse:
    properties:
      cards:
        items:
          $ref: '#/definitions/app.Card'
        type: array
      created_at:
        type: string
    type: object
info:
  contact: {}
  title: Swagger UI
  version: "1.0"
paths:
  /api/cards:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.Card'
            type: array
      summary: Получить массив карточек
      tags:
      - cards
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.CardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Card'
      summary: Создать карточку
      tags:
      - cards
  /api/cards/cart:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.Card'
            type: array
      summary: Получить массив карточек из корзины
      tags:
      - cart
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.Card'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Card'
      summary: добавить карточку в корзину
      tags:
      - cart
  /api/cards/cart/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: удалить карточку из корзины
      tags:
      - cart
  /api/cards/favorite:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.Card'
            type: array
      summary: Получить массив карточек из избранного
      tags:
      - favorite
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.Card'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Card'
      summary: добавить карточку в избранное
      tags:
      - favorite
  /api/cards/favorite/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: удалить карточку из избранного
      tags:
      - favorite
  /api/cards/order:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app.orderResponse'
            type: array
      summary: Получить массив карточек заказов
      tags:
      - order
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.orderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: добавить карточку в список заказов
      tags:
      - order
  /api/storage:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app.imageResponse'
      summary: Загрузить картинку
      tags:
      - storage
swagger: "2.0"